Dados recebidos de product.index() formatados incluindo categorias do produto
Correção de bugs e restrições adicionadas

+++ app/src/Controllers/ProductFactory.js
@@ -1,14 +1,50 @@
 const connection = require('../database/connection');
 
 const ProductController = require('./ProductController');
+const CategoryController = require('./CategoryController');
 const Product_CategoryController = require('./Product_CategoryController');
 
 module.exports = {
 
-    index(request, response){
+    async index(request, response){
 
-        const product = ProductController.index();
-        return response.json(product);
+        const products = await ProductController.index();
+
+        dataArray = new Array();
+
+        var i = 0;
+        var j = 0;
+        var k = 0;
+        var l = 0;
+
+        for (i ; products[i]!=undefined ; i++)
+        {
+            dataArray[k] = {
+                id: 0,
+                name: '',
+                description: '',
+                value: 0,
+                category: []
+            };
+            dataArray[k].id = products[i].id;
+            dataArray[k].name = products[i].name;
+            dataArray[k].description = products[i].description;
+            dataArray[k].value = products[i].value;
+            l = 0;
+            j = dataArray[k].id;
+
+            for (j ; products[i]!=undefined && j == products[i].id  ; i++)
+            { 
+                dataArray[k].category[l] = products[i].category;
+                l++;
+            }
+            i--;
+            k++;
+        }
+
+        console.log(dataArray);
+
+        return response.json(dataArray);
     },
 
     async createProduct(request, response){
@@ -27,16 +63,26 @@ module.exports = {
                 return response.status(400).send();
             }
 
-            try{
-                const product_id = await ProductController.create(name, description, value);
-                await Product_CategoryController.create(product_id, array);
-                console.log('Product create successful! ID: ', product_id);
-                return response.status(201).send();
-            }
-            catch{
-                console.log('\nError! Typed category does not exist or does not follow the model: id1,id2,id3...')
+            const categoryExist = await CategoryController.exist(array);
+
+            if(categoryExist == 0)
+            {
+                console.log('\nOne of the categories entered does not exist!')
                 console.log('\nProduct not created.');
                 return response.status(400).send();
+            }
+            else{
+                try{
+                    const product_id = await ProductController.create(name, description, value);
+                    await Product_CategoryController.create(product_id, array);
+                    console.log('\nProduct create successful! ID: ', product_id);
+                    return response.status(201).send();
+                }
+                catch{
+                    console.log('\nError! Typed category does not exist or does not follow the model: id1,id2,id3...')
+                    console.log('\nProduct not created.');
+                    return response.status(400).send();
+                }
             }
         } else
         {
@@ -46,13 +92,23 @@ module.exports = {
         }
     },
 
-    delete(request, response){
+    async delete(request, response){
         const { id } = request.params;
 
-        message = ProductController.delete(id);
-        Product_CategoryController.deleteProduct(id);
+        exist = await ProductController.verifyIfExistById(id);
+        if (exist == 1){
+            try{
+                await ProductController.delete(id);
+                await Product_CategoryController.deleteProduct(id);
+                console.log('\nProduct delete successfull!');
+            }
+            catch(error){
+                console.log(error);
+            }
+        }else{
+            console.log('\nProduct not found.');
+        }
 
-        console.log(message);
         return response.status(204).send();
     },
